Documentation: Update project title (#241)
* Update project title in README

Removed the ampersand from the project title for consistency and clarity.

* Update ERNIE API description in OpenAPI spec

Corrected the full form of ERNIE in the OpenAPI specification to 'Earth Research Notary for Information Editing' for improved accuracy.

* Update welcome page heading style and text

Changed the heading in the welcome page to use larger, bolder text and updated the title to 'ERNIE - Earth Research Notary for Information Editing'. Adjusted the related test to match the new heading text.

* Update About page text and styling

Revised the ERNIE acronym and description text, and changed paragraph styling to use larger, muted text for improved readability.

* Fix indentation in About page paragraph

Corrected the indentation of the first paragraph in the About page for improved code readability and consistency.